fix(oauth2): remove passport dependency
